feat: 최초 진입 이펙트 app으로 분리
