+++ frontend/src/App.tsx
@@ -1,3 +1,4 @@
+import { useContext, useEffect } from "react";
 import { BrowserRouter, Redirect, Route, Switch } from "react-router-dom";
 
 import { PAGE_URL } from "./constants/urls";
@@ -13,8 +14,34 @@ import UserFeedPage from "./pages/UserFeedPage/UserFeedPage";
 import TagFeedPage from "./pages/TagFeedPage/TagFeedPage";
 import SearchPage from "./pages/SearchPage/SearchPage";
 import SearchHeader from "./components/@layout/SearchHeader/SearchHeader";
+import UserContext from "./contexts/UserContext";
+import { getAccessToken } from "./storage/storage";
+import { requestGetSelfProfile } from "./services/requests";
+import SnackBarContext from "./contexts/SnackbarContext";
+import { SUCCESS_MESSAGE } from "./constants/messages";
 
 const App = () => {
+  const { currentUsername, login, logout } = useContext(UserContext);
+  const { pushSnackbarMessage } = useContext(SnackBarContext);
+
+  useEffect(() => {
+    const accessToken = getAccessToken();
+
+    if (!accessToken || !currentUsername) return;
+
+    (async () => {
+      try {
+        const { name } = await requestGetSelfProfile(accessToken);
+
+        login(accessToken, name);
+        pushSnackbarMessage(SUCCESS_MESSAGE.LOGIN);
+      } catch (error) {
+        logout();
+        pushSnackbarMessage(SUCCESS_MESSAGE.LOGOUT);
+      }
+    })();
+  }, []);
+
   return (
     <BrowserRouter>
       <Switch>
